refactor bridge pattern
