+++ Structural/Bridge/src/interfaces/NotificationDisplay.ts
@@ -1,0 +1,3 @@
+export interface NotificationDisplay {
+  display(message: string): void;
+}
