+++ Structural/Bridge/src/interfaces/SampleNotification.ts
@@ -1,0 +1,3 @@
+export interface SampleNotification {
+  send(message: string): void;
+}
